+++ ichnaea/api/locate/tests/test_internal.py
@@ -2,9 +2,10 @@ from ichnaea.api.locate.internal import InternalRegionSource
 from ichnaea.api.locate.tests.base import BaseSourceTest
 from ichnaea.geocode import GEOCODER
 from ichnaea.tests.factories import (
-    CellShardFactory,
+    CellAreaFactory,
     WifiShardFactory,
 )
+from ichnaea import util
 
 
 class TestRegionSource(BaseSourceTest):
@@ -14,11 +15,13 @@ class TestRegionSource(BaseSourceTest):
 
     def test_from_mcc(self):
         region = GEOCODER.regions_for_mcc(235, metadata=True)[0]
-        cell = CellShardFactory.build(mcc=235)
+        cell = CellAreaFactory(mcc=235, num_cells=10)
+        self.session.flush()
+
         query = self.model_query(cells=[cell])
         results = self.source.search(query)
         self.check_model_result(results, region)
-        self.assertAlmostEqual(results[0].score, 2.0, 4)
+        self.assertAlmostEqual(results[0].score, 1.0, 4)
         self.check_stats(counter=[
             (self.api_type + '.source',
                 ['key:test', 'region:none', 'source:internal',
@@ -26,13 +29,21 @@ class TestRegionSource(BaseSourceTest):
         ])
 
     def test_ambiguous_mcc(self):
+        now = util.utcnow()
         regions = GEOCODER.regions_for_mcc(234, metadata=True)
-        cell = CellShardFactory.build(mcc=234)
+        cell = CellAreaFactory(mcc=234, num_cells=10)
+        self.session.flush()
+
         query = self.model_query(cells=[cell])
         results = self.source.search(query)
         self.check_model_result(results, regions)
+        best_result = results.best(query.expected_accuracy)
+        self.assertEqual(best_result.region_code, 'GB')
         for result in results:
-            self.assertAlmostEqual(result.score, 0.5, 4)
+            score = 0.25
+            if result.region_code == 'GB':
+                score += cell.score(now)
+            self.assertAlmostEqual(result.score, score, 4)
         self.check_stats(counter=[
             (self.api_type + '.source',
                 ['key:test', 'region:none', 'source:internal',
@@ -40,25 +51,30 @@ class TestRegionSource(BaseSourceTest):
         ])
 
     def test_multiple_mcc(self):
+        now = util.utcnow()
         region = GEOCODER.regions_for_mcc(235, metadata=True)[0]
-        cell = CellShardFactory.build(mcc=234)
-        cell2 = CellShardFactory.build(mcc=235)
+        cell = CellAreaFactory(mcc=234, num_cells=6)
+        cell2 = CellAreaFactory(mcc=235, num_cells=8)
+        self.session.flush()
+
         query = self.model_query(cells=[cell, cell2])
         results = self.source.search(query)
         self.assertTrue(len(results) > 2)
         best_result = results.best(query.expected_accuracy)
         self.assertEqual(best_result.region_code, region.code)
-        self.assertAlmostEqual(best_result.score, 2.5, 4)
+        self.assertAlmostEqual(best_result.score, 1.25 + cell.score(now), 4)
 
     def test_invalid_mcc(self):
-        cell = CellShardFactory.build(mcc=235)
+        cell = CellAreaFactory.build(mcc=235, num_cells=10)
         cell.mcc = 999
         query = self.model_query(cells=[cell])
         results = self.source.search(query)
         self.check_model_result(results, None)
 
     def test_wifi(self):
-        wifis = WifiShardFactory.build_batch(2)
+        wifis = WifiShardFactory.create_batch(2)
+        self.session.flush()
+
         query = self.model_query(wifis=wifis)
         results = self.source.search(query)
         self.check_model_result(results, None)